layers: Remove more extra include headers
