+++ layers/best_practices/bp_drawdispatch.cpp
@@ -21,6 +21,7 @@
 #include "best_practices/best_practices_error_enums.h"
 #include "best_practices/bp_state.h"
 #include "state_tracker/buffer_state.h"
+#include "state_tracker/render_pass_state.h"
 #include <bitset>
 
 // Generic function to handle validation for all CmdDraw* type functions
