+++ layers/core_checks/cc_android.cpp
@@ -21,6 +21,7 @@
 #include <vulkan/vk_enum_string_helper.h>
 #include "core_validation.h"
 #include "state_tracker/image_state.h"
+#include "state_tracker/sampler_state.h"
 
 #if defined(VK_USE_PLATFORM_ANDROID_KHR)
 // Android-specific validation that uses types defined only on Android and only for NDK versions
