+++ layers/core_checks/cc_cmd_buffer.cpp
@@ -25,6 +25,8 @@
 #include "generated/chassis.h"
 #include "state_tracker/image_state.h"
 #include "state_tracker/buffer_state.h"
+#include "state_tracker/render_pass_state.h"
+#include "state_tracker/shader_module.h"
 #include "core_validation.h"
 #include "generated/enum_flag_bits.h"
