+++ layers/core_checks/cc_descriptor.cpp
@@ -22,10 +22,14 @@
 #include "state_tracker/descriptor_sets.h"
 #include "state_tracker/image_state.h"
 #include "state_tracker/buffer_state.h"
+#include "state_tracker/sampler_state.h"
+#include "state_tracker/render_pass_state.h"
 #include "state_tracker/ray_tracing_state.h"
+#include "state_tracker/shader_module.h"
 #include "cc_buffer_address.h"
 #include "generated/spirv_grammar_helper.h"
 #include "drawdispatch/descriptor_validator.h"
+#include "drawdispatch/drawdispatch_vuids.h"
 
 using DescriptorSet = vvl::DescriptorSet;
 using DescriptorSetLayout = vvl::DescriptorSetLayout;
