+++ layers/core_checks/cc_drawdispatch.cpp
@@ -23,6 +23,9 @@
 #include "state_tracker/image_state.h"
 #include "state_tracker/buffer_state.h"
 #include "state_tracker/shader_object_state.h"
+#include "state_tracker/descriptor_sets.h"
+#include "state_tracker/render_pass_state.h"
+#include "state_tracker/shader_module.h"
 
 using vvl::DrawDispatchVuid;
 using vvl::GetDrawDispatchVuid;
