+++ layers/core_checks/cc_spirv.cpp
@@ -29,6 +29,8 @@
 #include "utils/shader_utils.h"
 #include "utils/hash_util.h"
 #include "state_tracker/chassis_modification_state.h"
+#include "state_tracker/descriptor_sets.h"
+#include "state_tracker/render_pass_state.h"
 
 // Validate use of input attachments against subpass structure
 bool CoreChecks::ValidateShaderInputAttachment(const spirv::Module &module_state, const vvl::Pipeline &pipeline,
