+++ layers/gpu_validation/gpu_validation.cpp
@@ -32,6 +32,7 @@
 #include "containers/custom_containers.h"
 #include "spirv/module.h"
 #include "state_tracker/chassis_modification_state.h"
+#include "state_tracker/render_pass_state.h"
 // Generated shaders
 #include "generated/gpu_pre_draw_vert.h"
 #include "generated/gpu_pre_dispatch_comp.h"
