+++ layers/state_tracker/image_state.cpp
@@ -20,6 +20,7 @@
 #include "state_tracker/image_state.h"
 #include "state_tracker/pipeline_state.h"
 #include "state_tracker/descriptor_sets.h"
+#include "state_tracker/shader_module.h"
 #include <limits>
 #include <string_view>
