+++ layers/state_tracker/pipeline_sub_state.cpp
@@ -1,6 +1,6 @@
-/* Copyright (c) 2015-2017, 2019-2023 The Khronos Group Inc.
- * Copyright (c) 2015-2017, 2019-2023 Valve Corporation
- * Copyright (c) 2015-2017, 2019-2023 LunarG, Inc.
+/* Copyright (c) 2015-2017, 2019-2024 The Khronos Group Inc.
+ * Copyright (c) 2015-2017, 2019-2024 Valve Corporation
+ * Copyright (c) 2015-2017, 2019-2024 LunarG, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,6 +17,7 @@
 
 #include "state_tracker/pipeline_sub_state.h"
 #include "state_tracker/pipeline_state.h"
+#include "state_tracker/shader_module.h"
 
 VkPipelineLayoutCreateFlags PipelineSubState::PipelineLayoutCreateFlags() const {
     const auto layout_state = parent.PipelineLayoutState();
