+++ layers/state_tracker/pipeline_sub_state.h
@@ -1,6 +1,6 @@
-/* Copyright (c) 2015-2017, 2019-2023 The Khronos Group Inc.
- * Copyright (c) 2015-2017, 2019-2023 Valve Corporation
- * Copyright (c) 2015-2017, 2019-2023 LunarG, Inc.
+/* Copyright (c) 2015-2017, 2019-2024 The Khronos Group Inc.
+ * Copyright (c) 2015-2017, 2019-2024 Valve Corporation
+ * Copyright (c) 2015-2017, 2019-2024 LunarG, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -22,8 +22,11 @@
 
 // Graphics pipeline sub-state as defined by VK_KHR_graphics_pipeline_library
 
+class ValidationStateTracker;
 namespace vvl {
 class RenderPass;
+class Pipeline;
+class PipelineLayout;
 struct ShaderModule;
 }  // namespace vvl
