+++ layers/sync/sync_op.cpp
@@ -23,6 +23,7 @@
 
 #include "state_tracker/buffer_state.h"
 #include "state_tracker/cmd_buffer_state.h"
+#include "state_tracker/render_pass_state.h"
 
 #include "sync/sync_validation.h"
